Add images
bite the bullet, let big files into git
This is much the simplest way to manage them, and the only way that
really lets us update from multiple computers.
CDN should be the easiest way to serve them, too.

It's a hassle if we have to remove files from the repo; or switch to
git-annex someday, but the files shouldn't change, and shouldn't really
stress github.